backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.36
chromadb==0.4.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
langchain==0.1.9
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-community==0.0.27
langchain-google-genai==0.0.11
python-dotenv==1.0.1
PyPDF2==3.0.1
google-generativeai>=0.4.1,<0.5.0
firebase-admin>=6.0.0